==> Step-1 Project Definition.ipynb <==
<h1> Project Definition</h1>

<h3> Overview </h3>

Starbucks is a american coffeehouse company, which sends advertisement/offers to their customers via various channels. Out of every 100 adverstisement, there is 52% coversion rate to offer getting completed. So company is not able to reach to right audience via their adverstisent. Company wants to increase this conversion rate to 80%, so that company could 

* save their adverstisemnt cost.
* do cost-benefit analysis for an offer.
* Knowing customers helps in launching newer offer
* Save marketing cost to identify the right segment. 

<h3> Statement</h3>

Suppose if a offer is sent to customer, Is it possible to predict whether customer would complete offer or not? This right identification of audience can help us to optimize the market metrics. 

<h1> Metrics </h1>

In this predictive modelling exercise, Our objective is to know wheter customer would complete it or not. In this case, Our objective to achieve hight precision and high recall, so f1 socre is used by giving equal weight to both the objective. This is going to be our metrics of evaluation.
