==> Step-3 Methodology.ipynb <==
import pandas as pd

data = pd.concat([pd.read_csv("data_for_EDA.csv")])

del data['Unnamed: 0']

data['total reward per total amount'] = data['reward']/data['Total_amount']

data['Total_amount'] = data['Total_amount'].apply(lambda x : .00001 if x==0 else x)

data['time'] = data['time'].apply(lambda x : .0001 if x ==0 else x)

data['Total amount spent on website per month'] = data['Total_amount']*24*30/data['time']

data['Total reward got from website per month'] = data['reward']*24*30/data['time']

data['Total_offer_received'] = data['Total_offer_received'].apply(lambda x : .0001 if x ==0 else x)

data['Total viewed offer per Total sent offer'] = data['Total_view']/data['Total_offer_received']

data['Total completed offer per Total sent offer'] = data['offer completed']/data['Total_offer_received']

data['Total transaction made per month'] = data['total_transaction_made']*24*30/data['time']

data = data[['gender',
      'age',
      'income',
      'Year', #Year, when customer was acquired
      'month', #Year, when customer was acquired
      'web',
      'email', 
      'mobile', 
      'social',
     'offer_type',
      'duration',
      'difficulty',
      'reward',
      'week_day_in_which_offer_is_made',
      'total reward per total amount',
      'Total amount spent on website per month',
     'Total reward got from website per month',
      'Total viewed offer per Total sent offer',
     'Total completed offer per Total sent offer',
     'Total transaction made per month',
      'is_offer_completed'      
            ]]

for k in ['Year','month','week_day_in_which_offer_is_made']:
    data[k] = data[k].apply(lambda x : str(x))

import numpy as np

from sklearn.model_selection import train_test_split

from sklearn.ensemble import RandomForestClassifier

from sklearn.metrics import accuracy_score,roc_auc_score,confusion_matrix,precision_score,f1_score,recall_score

from imblearn.over_sampling import SMOTE

from sklearn.preprocessing import StandardScaler,MinMaxScaler,RobustScaler

import pandas as pd, numpy as np,pymysql,datetime as dt

from sklearn.feature_selection import VarianceThreshold

from imblearn.pipeline import make_pipeline

pipe = make_pipeline( RandomForestClassifier(random_state=100,n_estimators= 200) )

from sklearn.model_selection import GridSearchCV

gsc = GridSearchCV(   estimator=pipe,
    
                      param_grid={
                          
                          'randomforestclassifier__max_depth':[1,2,3,4,5,6,7,8,9,10],
                                  },
                      scoring='f1',
                      
                      cv= 5)

import warnings

warnings.filterwarnings("ignore")

data = pd.get_dummies(data)

data = data.fillna(data.median())

for k in data.columns:
    
    data[k] = data[k].apply(lambda x : 0 if x ==  np.inf else x)

columns = data.columns

from sklearn.preprocessing import RobustScaler,StandardScaler

scaler = StandardScaler()

data = pd.DataFrame(data)

data.columns = columns

y = data.is_offer_completed

X = data.drop("is_offer_completed",axis = 1)

data = scaler.fit_transform(data)

x_train, x_test, y_train, y_test = train_test_split(X,y, test_size = 0.2, random_state =1, stratify =y)

 gsc.fit(x_train, y_train)

y_pred = gsc.best_estimator_.predict(x_test)

f1_score(y_pred,y_test),precision_score(y_pred,y_test),recall_score(y_pred,y_test)

t = gsc.best_estimator_

max_dept = dict(t.steps)['randomforestclassifier'].max_depth

features = dict(t.steps)['randomforestclassifier'].feature_importances_

f = x_train.columns

from sklearn.feature_selection import SelectFromModel

sel_ = SelectFromModel(RandomForestClassifier(n_estimators=200,random_state=1,max_depth = max_dept),threshold="mean")

sel_.fit(x_train,y_train)

x_train.columns[(sel_.get_support())]

X_train_rf = pd.DataFrame(sel_.transform(x_train))
    
X_test_rf = pd.DataFrame(sel_.transform(x_test))

gsc2 = GridSearchCV(
        
    estimator=pipe,
    
    param_grid={

        
      'randomforestclassifier__max_depth':[1,2,3,4,5,6,7],
        
    },
    
    scoring='f1',
    cv=3 )
    
gsc2.fit(X_train_rf, y_train)
    
y_pred = gsc2.predict(X_test_rf)

f1_score(y_pred,y_test),precision_score(y_pred,y_test),recall_score(y_pred,y_test)

from sklearn.tree import DecisionTreeClassifier

pipe = make_pipeline( DecisionTreeClassifier(random_state=100))

gsc2 = GridSearchCV(
        
    estimator=pipe,
    
    param_grid={

        
      'decisiontreeclassifier__max_depth':[1,2,3,4,5,6,7,8,9,10],
        
    },
    
    scoring='f1',
    cv=3 )
    
gsc2.fit(X_train_rf, y_train)
    
y_pred = gsc2.predict(X_test_rf)

f1_score(y_pred,y_test),precision_score(y_pred,y_test),recall_score(y_pred,y_test),accuracy_score(y_pred,y_test)

from sklearn.linear_model import LogisticRegression

pipe = make_pipeline( LogisticRegression(random_state=100))
gsc2 = GridSearchCV(
        
    estimator=pipe,
    
    param_grid={

        
      'logisticregression__penalty':['l2','l1'],
        
    },
    
    scoring='f1',
    
    cv=3 
                  )
    
gsc2.fit(X_train_rf, y_train)
    
y_pred = gsc2.predict(X_test_rf)

f1_score(y_pred,y_test),precision_score(y_pred,y_test),recall_score(y_pred,y_test),accuracy_score(y_pred,y_test)

from sklearn.svm import SVC

svm = SVC(random_state=1)

svm.fit(X_train_rf, y_train)
    
y_pred = svm.predict(X_test_rf)

f1_score(y_pred,y_test),precision_score(y_pred,y_test),recall_score(y_pred,y_test),accuracy_score(y_pred,y_test)

f1_score(y_pred,y_test),precision_score(y_pred,y_test),recall_score(y_pred,y_test),accuracy_score(y_pred,y_test)
